@@ -17,7 +17,7 @@ import { LongReadsubmittedSequencesService } from './api/longReadsubmittedSequen
     LongReadsubmittedSequencesService ]
 })
 export class ApiModule {
-    public static forRoot(configurationFactory: () => Configuration): ModuleWithProviders {
+    public static forRoot(configurationFactory: () => Configuration): ModuleWithProviders<ApiModule> {
         return {
             ngModule: ApiModule,
             providers: [ { provide: Configuration, useFactory: configurationFactory } ]
